限制检查器 长度检查

+++ NumberChecker.cpp
@@ -5,8 +5,8 @@ NumberChecker::NumberChecker(const Section& config) {
 		std::string rangeStr = config.at("Range").value;
 		size_t commaPos = rangeStr.find(',');
 		if (commaPos != std::string::npos) {
-			minRange = std::stoi(rangeStr.substr(0, commaPos));
-			maxRange = std::stoi(rangeStr.substr(commaPos + 1));
+			minRange = std::stof(rangeStr.substr(0, commaPos));
+			maxRange = std::stof(rangeStr.substr(commaPos + 1));
 		}
 	}
 
@@ -16,7 +16,7 @@ NumberChecker::NumberChecker(const Section& config) {
 }
 
 void NumberChecker::validate(const std::string& value) const {
-	int intValue = std::stoi(value);
+	float intValue = std::stof(value);
 	if (!checkRange(intValue))
 		throw "值 " + value + " 不在范围 [" + std::to_string(minRange) + ", " + std::to_string(maxRange) + "] 内";
 }
